Removes PlayerMovementAlt.cs
This was an old, unused file that implemented a slightly different
movement system that allowed players to move while using abilities. It
is no longer necessary to keep as the player scripts have developed to
the point where it is more efficient to make the change in existing code
than try to update this file.